Header: fix hydration error, that was causing light mode reset
NavigationMenuLink renders an <a> element, and if it's child has
<Link> component (also renders <a>), it will cause a hydration
error as you can't have nested <a> components.

The fix is to move <Link> inside of the
NavigationMenuLink and pass asChild to it. This will make it so
the default DOM element of NavigationMenuLink will not be rendered,
instead the child component (<Link>) as <a> will be rendered and it
will inherit the behaviour.

Inspired by: https://stackoverflow.com/questions/77944773/next-js-14-shadcn-error-hydration-failed-because-the-initial-ui-does-not-matc

Signed-off-by: John Titor <50095635+JohnRTitor@users.noreply.github.com>